+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJCSPZTAKSH7RE5RMHPVA1SX"
+        "01KJCYYPF58EQMZW3S0AYEXMDC"
     )
     r"""The unique identifier of the data source"""