+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -259,7 +259,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KJCSPZPE6B6RHWYGFQJH726S"
+    id: Optional[str] = "01KJCYYPASJSHKSD005YD2VQ0Q"
 
     description: Optional[str] = None
 
@@ -1214,7 +1214,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJCSPZP0RM51K0QT6EPB5WFX"
+        "tool_01KJCYYPAGWWB2QJ3HSH0TAE15"
     )
 
     display_name: Optional[str] = None
@@ -1345,7 +1345,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KJCSPZNZBTMTHZS04RYV27AH"
+    id: Optional[str] = "01KJCYYPAE1VPVSHZPDRX23N8Q"
 
     description: Optional[str] = None
 
@@ -1476,7 +1476,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJCSPZNYTCK2KSQEVZDFX5R3"
+        "tool_01KJCYYPAC8ARKNTZKGEAYDZCG"
     )
 
     display_name: Optional[str] = None
@@ -1777,7 +1777,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJCSPZNWAZE7A3R6RFWE40HC"
+        "tool_01KJCYYPA9EAFA1PVW76WY4Y9Y"
     )
 
     display_name: Optional[str] = None
@@ -1972,7 +1972,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJCSPZNV3M7RGTPCHVM0R3B3"
+        "tool_01KJCYYPA7G2H8KYVJ91J0CDH8"
     )
 
     display_name: Optional[str] = None
@@ -2171,7 +2171,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJCSPZNSQ9GH39CJER1QR51F"
+        "tool_01KJCYYPA554R4J4EBBFVHSDQQ"
     )
 
     display_name: Optional[str] = None